requirements.txt
aiopg==1.3.3
asgiref==3.4.1
async-timeout==4.0.1
backports.zoneinfo==0.2.1
Django==3.2.9
django-cors-headers==3.10.1
djangorestframework==3.12.4
gunicorn==20.1.0
pkg_resources==0.0.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
whitenoise==5.3.0
psycopg2==2.8.5

